+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     IDs if `cache_build_ids` is `true`
 - Added support for compressed debug information
   - Added `zlib` (default enabled) and `zstd` (default disabled) features
+- Added support for transparently following debug links in ELF binaries
 - Adjusted `Inspector::for_each` signature to no longer carry explicit state
   around
 - Introduced `normalize::Reason` enum to provide best guess at why normalization